Now it supports adding and deleting

+++ Telemaco/src/com/diegomartin/telemaco/control/TripControl.java
@@ -34,7 +34,7 @@ public class TripControl {
 		return TripDAO.read();
 	}
 	
-	public static void deleteTrip(int id){
+	public static void deleteTrip(long id){
 		TripDAO.delete(id);
 	}
 }
